requirements.txt
python-rapidjson
psycopg2
sqlalchemy
